+++ src/main/java/ru/vyarus/guice/persist/orient/repository/command/query/Query.java
@@ -46,5 +46,5 @@ public @interface Query {
      *
      * @return connection hint
      */
-    DbType connection() default DbType.DOCUMENT;
+    DbType connection() default DbType.UNKNOWN;
 }