+++ src/main/java/ru/vyarus/guice/persist/orient/repository/command/script/Script.java
@@ -61,5 +61,5 @@ public @interface Script {
      *
      * @return connection hint
      */
-    DbType connection() default DbType.DOCUMENT;
+    DbType connection() default DbType.UNKNOWN;
 }