+++ src/main/java/ru/vyarus/guice/persist/orient/repository/core/executor/ExecutorAnalyzer.java
@@ -1,12 +1,8 @@
 package ru.vyarus.guice.persist.orient.repository.core.executor;
 
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 import ru.vyarus.guice.persist.orient.db.DbType;
 import ru.vyarus.guice.persist.orient.repository.core.result.ResultDescriptor;
-import ru.vyarus.guice.persist.orient.repository.core.util.RepositoryUtils;
 
-import java.lang.reflect.Method;
 import java.util.Set;
 
 import static ru.vyarus.guice.persist.orient.repository.core.MethodDefinitionException.check;
@@ -18,43 +14,49 @@ import static ru.vyarus.guice.persist.orient.repository.core.MethodDefinitionExc
  * @since 26.09.2014
  */
 public final class ExecutorAnalyzer {
-    private static final Logger LOGGER = LoggerFactory.getLogger(ExecutorAnalyzer.class);
 
     private ExecutorAnalyzer() {
     }
 
     /**
-     * Selects appropriate executor for repository method.
+     * Selects appropriate executor for repository method. Custom converters most likely will cause
+     * method return type different from raw object, returned from connection. So in such case
+     * detection of connection from return type is impossible.
+     * <p>If custom converter registered: always use connection hint if available. Note that result
+     * converter could also change connection hint.</p>
+     * <p>If no custom converter register then if connection hint contradict with result type
+     * analysis throw an error.</p>
      *
-     * @param method          method to analyze
-     * @param descriptor      result definition
-     * @param executors       available executors
-     * @param defaultExecutor default executor to use
-     * @param connectionHint  expected connection type hint
+     * @param descriptor          result definition
+     * @param executors           available executors
+     * @param defaultExecutor     default executor to use
+     * @param connectionHint      expected connection type hint
+     * @param customConverterUsed true when custom converter registered
      * @return selected executor instance
      */
     public static RepositoryExecutor analyzeExecutor(
-            final Method method,
             final ResultDescriptor descriptor,
             final Set<RepositoryExecutor> executors,
             final RepositoryExecutor defaultExecutor,
-            final DbType connectionHint) {
+            final DbType connectionHint,
+            final boolean customConverterUsed) {
 
-        // even if connection hint available trying to detect to later check compatibility
-        RepositoryExecutor executor = selectByType(descriptor.entityType, executors);
-        final DbType requestedType = getRequestedConnectionType(method, executor, connectionHint);
-
-
-        if (executor == null) {
-            executor = requestedType != null ? find(requestedType, executors) : defaultExecutor;
+        RepositoryExecutor executor;
+        if (connectionHint != null) {
+            // connection hint in priority
+            executor = find(connectionHint, executors);
+            check(executor != null, "Executor not found for type set in annotation %s", connectionHint);
+            if (!customConverterUsed) {
+                // catch silly errors
+                validateHint(selectByType(descriptor.entityType, executors), connectionHint);
+            }
         } else {
-            // special case, sometimes document connection could be overridden, for example:
-            // when querying for fields in object connection, documents returned, but still we can use object connection
-            if (executor.getType().equals(DbType.DOCUMENT) && requestedType != null) {
-                executor = find(requestedType, executors);
+            // automatic detection
+            executor = selectByType(descriptor.entityType, executors);
+            if (executor == null) {
+                executor = defaultExecutor;
             }
         }
-        check(executor != null, "Executor not found for type set in annotation %s", requestedType);
         return executor;
     }
 
@@ -69,20 +71,17 @@ public final class ExecutorAnalyzer {
         return executor;
     }
 
-    private static DbType getRequestedConnectionType(final Method method, final RepositoryExecutor executor,
-                                                     final DbType connectionHint) {
-        // ignore document as default value
-        final DbType requestedType = connectionHint == DbType.DOCUMENT ? null : connectionHint;
-
-        // annotation guides just ambiguous cases
-        if (executor != null && requestedType != null
-                && !executor.getType().equals(requestedType)) {
-            LOGGER.warn(
-                    "Connection hint {} ignored, because correct execution type recognized from return type "
-                            + "in repository method {}",
-                    connectionHint, RepositoryUtils.methodToString(method));
+    private static void validateHint(final RepositoryExecutor executor, final DbType connectionHint) {
+        if (executor != null) {
+            final DbType autoType = executor.getType();
+            if (autoType.equals(DbType.DOCUMENT) && connectionHint.equals(DbType.OBJECT)) {
+                // it's ok to use object connection for document selection
+                return;
+            }
+            check(autoType.equals(connectionHint),
+                    "Bad connection hint %s specified, when %s expected (according to return type).",
+                    connectionHint, autoType);
         }
-        return requestedType;
     }
 
     private static RepositoryExecutor find(final DbType type, final Set<RepositoryExecutor> executors) {