+++ src/main/java/ru/vyarus/guice/persist/orient/repository/delegate/Delegate.java
@@ -78,5 +78,5 @@ public @interface Delegate {
      *
      * @return connection hint
      */
-    DbType connection() default DbType.DOCUMENT;
+    DbType connection() default DbType.UNKNOWN;
 }